Queue overhaul, no-winner bug fix, style
Queue changed to class extending py list, other functions removed/wraped
into it. WANT_FILL only referenced inside the class, and took
complicated manipulations around the logic to functions of the class.

If a game was exit'd, the replay on the site would not display a winner,
resulting in a critical error.

my choice of style changes around the document.